add types to single folder
