+++ src/api/redux/api/card-api.ts
@@ -1,51 +1,8 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import {
-  CharecterCard,
-  Gender,
-  Status,
-} from "../../../shared/components/Card/Card";
+import { CharecterCard } from "../../../shared/components/Card/Card";
 import parseLoadedCharecters from "../../../shared/utils/parseLoadedCharecters";
 import parseSingleLoadedCharecter from "../../../shared/utils/parseSingleLoadedCharacter";
-
-interface Location {
-  name: string;
-  url: string;
-}
-
-interface Origin {
-  name: string;
-  url: string;
-}
-
-export interface Result {
-  createdAt: string;
-  episode: string[] | string;
-  gender: Gender;
-  id: number;
-  image: string;
-  location: Location;
-  name: string;
-  origin: Origin;
-  species: string;
-  status: Status;
-  type: string;
-  url: string;
-}
-
-interface ResponseType {
-  info: {
-    count: number;
-    next: string | null;
-    pages: number;
-    prev: string | null;
-  };
-  results: Result[];
-}
-
-interface TransfromedResponse {
-  hasNextPage: String | null;
-  cards: CharecterCard[];
-}
+import { ResponseType, Result, TransfromedResponse } from "../../types";
 
 export const cardApi = createApi({
   reducerPath: "cardApi",
