--- src/shared/hooks/index.ts
@@ -1,1 +1,0 @@
-export const tmp = 0;
